pyproject.toml
[tool.poetry]
name = "goodscoop"
version = "0.1.0"
description = "A personalized news summary bot for Telegram"
authors = ["thatgardnerone"]
packages = [
    { include = "app" },
    { include = "config" },
    { include = "main.py" }
]

[tool.poetry.dependencies]
python = "^3.12"
ollama = "^0.4.6"
python-telegram-bot = "^21.10"
APScheduler = "^3.11.0"
pygooglenews = "^0.1.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
markdownify = "^0.14.1"
httpx = "^0.27.2"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"
feedparser = "^6.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.8"

[tool.poetry.scripts]
goodscoop = "main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
